+++ CHANGELOG.md
@@ -30,7 +30,7 @@
 - Added jruby support in general and in Travis
 - #196 - Topic mapper does not map topics when subscribing thanks to @webandtech
 - #96 - Karafka server - possiblity to run it only for a certain topics
-- karafka worker cli option is removed (please use sidekiq directly)
+- ~~karafka worker cli option is removed (please use sidekiq directly)~~ - restored, bad idea
 - (optional) pausing upon processing failures ```pause_timeout```
 
 ### New features and improvements